Public API

Contents

Public API

Contents

Index

Public interface

Index

Crystallography.SeitzOperator

CrystallographyBase.CartesianFromFractional

CrystallographyBase.FractionalFromCartesian

CrystallographyBase.MetricTensor

CrystallographyBase.PrimitiveFromStandardized

CrystallographyBase.ReciprocalLattice

CrystallographyBase.ReciprocalPoint

CrystallographyBase.StandardizedFromPrimitive

CrystallographyCore.Lattice

CrystallographyBase.cellvolume

CrystallographyBase.coordinates

CrystallographyBase.distance

CrystallographyBase.supercell

CrystallographyBase.weights

Public interface

Lattice

Missing docstring.

Missing docstring for CrystalSystem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Triclinic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Monoclinic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Orthorhombic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Tetragonal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Cubic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Trigonal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Hexagonal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Centering. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BaseCentering. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Primitive. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BodyCentering. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FaceCentering. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RhombohedralCentering. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BaseCentering. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Bravais. Check Documenter's build log for details.

CrystallographyCore.Lattice — Type

Lattice(data::AbstractMatrix)

Construct a Lattice from a matrix.

Note

The basis vectors of the matrix are stored as columns.

Examples

julia> Lattice([
    1.2 4.5 7.8
    2.3 5.6 8.9
    3.4 6.7 9.1
])
3×3 Lattice{Float64}
 1.2  4.5  7.8
 2.3  5.6  8.9
 3.4  6.7  9.1

Lattice(𝐚::AbstractVector, 𝐛::AbstractVector, 𝐜::AbstractVector)

Construct a Lattice from three basis vectors.

Examples

julia> 𝐚, 𝐛, 𝐜 = [1.2, 2.3, 3.4], [4.5, 5.6, 6.7], [7.8, 8.9, 9.10];

julia> Lattice(𝐚, 𝐛, 𝐜)
3×3 Lattice{Float64}
 1.2  4.5  7.8
 2.3  5.6  8.9
 3.4  6.7  9.1

Lattice(data)

Construct a Lattice from, e.g., a vector of three basis vectors.

Examples

julia> Lattice([[1.2, 2.3, 3.4], [4.5, 5.6, 6.7], [7.8, 8.9, 9.10]])
3×3 Lattice{Float64}
 1.2  4.5  7.8
 2.3  5.6  8.9
 3.4  6.7  9.1

julia> Lattice(((1.1, 2.2, 3.1), (4.4, 5.5, 6.5), (7.3, 8.8, 9.9)))
3×3 Lattice{Float64}
 1.1  4.4  7.3
 2.2  5.5  8.8
 3.1  6.5  9.9

julia> Lattice((1.1, 2.2, 3.1, 4.4, 5.5, 6.5, 7.3, 8.8, 9.9))
3×3 Lattice{Float64}
 1.1  4.4  7.3
 2.2  5.5  8.8
 3.1  6.5  9.9

julia> Lattice(i * 1.1 for i in 1:9)
3×3 Lattice{Float64}
 1.1  4.4  7.700000000000001
 2.2  5.5  8.8
 3.3000000000000003  6.6000000000000005  9.9

julia> using Unitful, UnitfulAtomic

julia> Lattice([
    [4u"nm", 0u"m", 0.0u"cm"],
    [0u"cm", 180.0u"bohr", 0u"m"],
    [0u"bohr", 0u"nm", (3//1) * u"angstrom"],
])
3×3 Lattice{Quantity{Float64, 𝐋, Unitful.FreeUnits{(m,), 𝐋, nothing}}}
 4.0e-9 m  0.0 m  0.0 m
 0.0 m  9.525189796254e-9 m  0.0 m
 0.0 m  0.0 m  3.0e-10 m

Lattice(cell::Cell)

Get the lattice of a Cell.

Lattice(a, b, c, α, β, γ; axis = :a)

Construct a Lattice from the six cell parameters.

The default convention we used here is that edge vector 𝐚 in the positive x-axis direction, edge vector 𝐛 in the x-y plane with a positive y-axis component, and edge vector 𝐜 with a positive z-axis component in the Cartesian system. See Wikipedia. You can also choose axis = :c.

Lattice(g::MetricTensor)

Construct a Lattice from a MetricTensor.

Missing docstring.

Missing docstring for centering. Check Documenter's build log for details.

Missing docstring.

Missing docstring for crystalsystem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for basis_vectors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for cellparameters. Check Documenter's build log for details.

CrystallographyBase.supercell — Function

supercell(lattice::Lattice, repfactors::AbstractMatrix{<:Integer})
supercell(lattice::Lattice, repfactors::AbstractVector{<:Integer})
supercell(lattice::Lattice, repfactor::Integer)

Create a supercell from lattice.

supercell(cell::Cell, repfactors::AbstractMatrix{<:Integer})
supercell(cell::Cell, repfactors::AbstractVector{<:Integer})
supercell(cell::Cell, repfactor::Integer)

Create a supercell from cell.

Note

Currently, only integral replications are supported.

Reciprocal space

Note that we take $2pi$ as $1$, not the solid-state physics convention.

CrystallographyBase.ReciprocalPoint — Type

ReciprocalPoint(x, y, z, w)

Represent a special point of the 3D Brillouin zone. Each of them has a weight w.

CrystallographyBase.ReciprocalLattice — Type

ReciprocalLattice(data::MMatrix)

Construct a ReciprocalLattice.

Warning

You should not use this function directly, always use reciprocal of a Lattice.

Missing docstring.

Missing docstring for inv. Check Documenter's build log for details.

Missing docstring.

Missing docstring for reciprocal_mesh. Check Documenter's build log for details.

CrystallographyBase.coordinates — Function

coordinates(arr::AbstractArray{<:ReciprocalPoint})

Get the coordinates of an array of ReciprocalPoints.

CrystallographyBase.weights — Function

weights(arr::AbstractArray{<:ReciprocalPoint})

Get the weights of an array of ReciprocalPoints.

Miller and Miller–Bravais indices

Missing docstring.

Missing docstring for Miller. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MillerBravais. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ReciprocalMiller. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ReciprocalMillerBravais. Check Documenter's build log for details.

Missing docstring.

Missing docstring for family. Check Documenter's build log for details.

Missing docstring.

Missing docstring for @m_str. Check Documenter's build log for details.

Metric tensor

CrystallographyBase.MetricTensor — Type

MetricTensor(𝐚::AbstractVector, 𝐛::AbstractVector, 𝐜::AbstractVector)

Generate a MetricTensor from the three basis vectors.

MetricTensor(lattice::Lattice)

Generate a MetricTensor from a Lattice.

MetricTensor(a, b, c, α, β, γ)

Generate a MetricTensor from the six cell parameters.

Missing docstring.

Missing docstring for directioncosine. Check Documenter's build log for details.

Missing docstring.

Missing docstring for directionangle. Check Documenter's build log for details.

CrystallographyBase.distance — Function

distance(𝐚::AbstractVector, g::MetricTensor, 𝐛::AbstractVector)

Get the distance between two coordinates using a MetricTensor.

Missing docstring.

Missing docstring for interplanar_spacing. Check Documenter's build log for details.

Transformations

CrystallographyBase.CartesianFromFractional — Type

CartesianFromFractional(lattice::Union{Lattice,ReciprocalLattice})
CartesianFromFractional(a, b, c, α, β, γ)

Get the transformation from fractional coordinates to Cartesian coordinates.

CrystallographyBase.FractionalFromCartesian — Type

FractionalFromCartesian(lattice::Union{Lattice,ReciprocalLattice})
FractionalFromCartesian(a, b, c, α, β, γ)

Get the transformation from Cartesian coordinates to fractional coordinates.

CrystallographyBase.PrimitiveFromStandardized — Type

PrimitiveFromStandardized(tf::AbstractMatrix)

Construct the transformation from a standardized cell to a primitive cell.

CrystallographyBase.StandardizedFromPrimitive — Type

StandardizedFromPrimitive(tf::AbstractMatrix)

Construct the transformation from a primitive cell to a standardized cell.

Others

CrystallographyBase.cellvolume — Function

cellvolume(a, b, c, α, β, γ)

Calculate the cell volume from 6 cell parameters.

cellvolume(l::Lattice)
cellvolume(c::Cell)

Calculate the cell volume from a Lattice or a Cell.

cellvolume(g::MetricTensor)

Calculate the cell volume from a MetricTensor.

Symmetry

Crystallography.SeitzOperator — Type

SeitzOperator(𝐑::AbstractMatrix, 𝐭::AbstractVector)
SeitzOperator(𝐑::AbstractMatrix)
SeitzOperator(𝐭::AbstractVector)

Construct a Seitz operator from matrices, vectors, or both.

The operator is defined by the following equation:

\[\left[ \begin{array}{ccc|c} & & & \\ & \mathbf{R} & & \mathbf{t} \\ & & & \\ \hline 0 & 0 & 0 & 1 \end{array} \right],\]

where $\mathbf{R}$ is a point group operation and $\mathbf{t}$ is a translation.